<==== Mercurial data ====>
Branch: vms_4.0_server

+++ vms/libs/nx_sdk/src/nx/sdk/analytics/helpers/event_metadata_packet.cpp
@@ -34,12 +34,12 @@ int64_t EventMetadataPacket::durationUs() const
 
 int EventMetadataPacket::count() const
 {
-    return m_events.size();
+    return (int) m_events.size();
 }
 
 const IEventMetadata* EventMetadataPacket::at(int index) const
 {
-    if (index < 0 || index >= m_events.size())
+    if (index < 0 || index >= (int) m_events.size())
         return nullptr;
 
     return m_events[index];
